+++ framework/Source/CPTMutableTextStyle.h
@@ -9,5 +9,6 @@
 @property (readwrite, assign, nonatomic) CGFloat fontSize;
 @property (readwrite, copy, nonatomic) CPTColor *color;
 @property (readwrite, assign, nonatomic) CPTTextAlignment textAlignment;
+@property (readwrite, assign, nonatomic) NSLineBreakMode lineBreakMode;
 
 @end