+++ framework/Source/CPTMutableTextStyle.m
@@ -27,4 +27,9 @@
  **/
 @dynamic textAlignment;
 
+/** @property NSLineBreakMode lineBreakMode
+ *  @brief The line break mode used when laying out the text. Default is #NSLineBreakByWordWrapping.
+ **/
+@dynamic lineBreakMode;
+
 @end