+++ framework/Source/CPTTextStyle.m
@@ -12,6 +12,7 @@
 @property (readwrite, assign, nonatomic) CGFloat fontSize;
 @property (readwrite, copy, nonatomic) CPTColor *color;
 @property (readwrite, assign, nonatomic) CPTTextAlignment textAlignment;
+@property (readwrite, assign, nonatomic) NSLineBreakMode lineBreakMode;
 
 @end
 
@@ -46,6 +47,11 @@
  **/
 @synthesize textAlignment;
 
+/** @property NSLineBreakMode lineBreakMode
+ *  @brief The line break mode used when laying out the text. Default is #NSLineBreakByWordWrapping.
+ **/
+@synthesize lineBreakMode;
+
 #pragma mark -
 #pragma mark Factory Methods
 
@@ -70,6 +76,7 @@
  *  - @ref fontSize = @num{12.0}
  *  - @ref color = opaque black
  *  - @ref textAlignment = #CPTTextAlignmentLeft
+ *  - @ref lineBreakMode = #NSLineBreakByWordWrapping
  *
  *  @return The initialized object.
  **/
@@ -80,6 +87,7 @@
         fontSize      = CPTFloat(12.0);
         color         = [[CPTColor blackColor] retain];
         textAlignment = CPTTextAlignmentLeft;
+        lineBreakMode = NSLineBreakByWordWrapping;
     }
     return self;
 }
@@ -108,6 +116,7 @@
     [coder encodeCGFloat:self.fontSize forKey:@"CPTTextStyle.fontSize"];
     [coder encodeObject:self.color forKey:@"CPTTextStyle.color"];
     [coder encodeInt:self.textAlignment forKey:@"CPTTextStyle.textAlignment"];
+    [coder encodeInteger:(NSInteger)self.lineBreakMode forKey:@"CPTTextStyle.lineBreakMode"];
 }
 
 -(id)initWithCoder:(NSCoder *)coder
@@ -117,6 +126,7 @@
         self->fontSize      = [coder decodeCGFloatForKey:@"CPTTextStyle.fontSize"];
         self->color         = [[coder decodeObjectForKey:@"CPTTextStyle.color"] copy];
         self->textAlignment = (CPTTextAlignment)[coder decodeIntForKey : @"CPTTextStyle.textAlignment"];
+        self->lineBreakMode = (NSLineBreakMode)[coder decodeIntegerForKey : @"CPTTextStyle.lineBreakMode"];
     }
     return self;
 }
@@ -136,6 +146,7 @@
     newCopy->color         = [self->color copy];
     newCopy->fontSize      = self->fontSize;
     newCopy->textAlignment = self->textAlignment;
+    newCopy->lineBreakMode = self->lineBreakMode;
     return newCopy;
 }
 
@@ -154,6 +165,7 @@
     newCopy->color         = [self->color copy];
     newCopy->fontSize      = self->fontSize;
     newCopy->textAlignment = self->textAlignment;
+    newCopy->lineBreakMode = self->lineBreakMode;
     return newCopy;
 }