+++ documentation/changelog.markdown
@@ -10,6 +10,7 @@ To be determined.
 - **New**: Added a property to hide plot data labels.
 - **New**: Added support for momentum scrolling.
 - **New**: Added support for "rubber band" snap-back when scrolling reaches the global x- and y-ranges.
+- **New**: Added line break mode support to `CPTTextStyle`.
 - **Changed**: Miscellaneous bug fixes and cleanup.